+++ src/main/typescript/constants.ts
@@ -129,6 +129,9 @@ export function isPhysicalItemType(
   return PhysicalItemTypes.includes(type as PhysicalItemType);
 }
 
+export type KnownTag = typeof KnownTags[number];
+export const KnownTags = ["skillful"];
+
 export type Rarity = typeof Rarities[number];
 export const Rarities = ["common", "uncommon", "rare", "exotic"] as const;